PTML: move location creation to the type builder
